refactor events serializers
