+++ events/Email.py
@@ -2,8 +2,6 @@ from django.core.mail import EmailMultiAlternatives
 from django.template.loader import render_to_string
 from django.utils.html import strip_tags
 
-
-
 def send_ticket_email(registration):
     ticket_details = {
         'ticket_number': str(registration.ticket_number),
@@ -28,22 +26,3 @@ def send_ticket_email(registration):
     email.attach_alternative(html_message, "text/html")
     email.send(fail_silently=False)
 
-
-def send_the_otp_email(user, otp):
-    otp_details = {
-        'user': user.first_name,
-        'otp': otp.otp_code
-    }
-    subject = 'Verify Your Email Address'
-    html_message = render_to_string('OTP/otp.html', {'otp_details': otp_details})
-    plain_message = strip_tags(html_message)
-
-    email = EmailMultiAlternatives(
-        subject,
-        plain_message,
-        'ondeyostephen0@gmail.com',
-
-        [user.email]
-    )
-    email.attach_alternative(html_message, "text/html")
-    email.send(fail_silently=False)
